Refactor dashboard
Move sonification and d3 logic to their own files,
makes dashbord.tsx a react functional component.
